fix-#3 [타이머][UI] 하단 숫자 버튼 추상화
- 하단 숫자 버튼 공통적으로 상용할 수 있게 변경
