+++ app/src/main/java/cookcook/nexters/com/amoogye/views/timer/presenter/TimerFragment.kt
@@ -9,6 +9,7 @@ import android.widget.RelativeLayout
 import androidx.databinding.DataBindingUtil
 import cookcook.nexters.com.amoogye.R
 import cookcook.nexters.com.amoogye.base.BaseFragment
+import cookcook.nexters.com.amoogye.base.BaseNumberButton
 import cookcook.nexters.com.amoogye.databinding.FragmentTimerBinding
 import kotlinx.android.synthetic.main.fragment_timer.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -22,36 +23,6 @@ class TimerFragment : BaseFragment() {
 
     override val isUseDataBinding: Boolean = true
 
-    private var numberButtonList = arrayOf(
-        R.id.layout_btn_0,
-        R.id.layout_btn_1,
-        R.id.layout_btn_2,
-        R.id.layout_btn_3,
-        R.id.layout_btn_4,
-        R.id.layout_btn_5,
-        R.id.layout_btn_6,
-        R.id.layout_btn_7,
-        R.id.layout_btn_8,
-        R.id.layout_btn_9,
-        R.id.layout_btn_dot,
-        R.id.layout_btn_delete
-    )
-
-    private var numberButtonTextList = arrayOf(
-        "0",
-        "1",
-        "2",
-        "3",
-        "4",
-        "5",
-        "6",
-        "7",
-        "8",
-        "9",
-        ".",
-        "delete"
-    )
-
     override fun setupViews(view: View) {
         timerViewModel.timerEditTextSetting(edit_hour)
         timerViewModel.timerEditTextSetting(edit_minute)
@@ -61,22 +32,23 @@ class TimerFragment : BaseFragment() {
         editTextClickEvent(edit_minute, 1)
         editTextClickEvent(edit_second, 2)
 
-        for (x in 0 until numberButtonList.size) {
-            var layout = view.findViewById<RelativeLayout>(numberButtonList[x])
-            layout.setOnClickListener {
-                timerViewModel.onNumberButtonClick(numberButtonTextList[x])
-            }
-        }
+        BaseNumberButton(view, onClick)
+    }
+
+    private val onClick: (number: String) -> Unit = {
+        timerViewModel.onNumberButtonClick(it)
     }
 
     private fun editTextClickEvent(editText: EditText, number: Int) {
-        editText.setOnTouchListener { v, event ->   event.let {
-            if(it.action == MotionEvent.ACTION_DOWN) {
-                timerViewModel.setSelectedTimerEditText(number)
+        editText.setOnTouchListener { v, event ->
+            event.let {
+                if (it.action == MotionEvent.ACTION_DOWN) {
+                    timerViewModel.setSelectedTimerEditText(number)
+                }
+                editText.setSelection(editText.text.length)
+                false
             }
-            editText.setSelection(editText.text.length)
-            false
-        }}
+        }
     }
 
     override fun subscribeUI() {
