+++ app/src/main/res/layout/fragment_timer.xml
@@ -28,8 +28,8 @@
                         android:id="@+id/edit_hour"
                         android:padding="4dp"
                         android:layout_width="30dp"
+                        android:cursorVisible="false"
                         android:text="@{timerVM.hourText}"
-                        android:selection="@{timerVM.hourText.length}"
                         android:background="@drawable/number_input_wrap_rounded_box"
                         android:layout_height="30dp"/>
 
@@ -44,8 +44,8 @@
                         android:id="@+id/edit_minute"
                         android:padding="4dp"
                         android:layout_width="30dp"
+                        android:cursorVisible="false"
                         android:text="@{timerVM.minuteText}"
-                        android:selection="@{timerVM.minuteText.length}"
                         android:background="@drawable/number_input_wrap_rounded_box"
                         android:layout_marginLeft="10dp"
                         android:layout_height="30dp"/>
@@ -60,9 +60,9 @@
                 <EditText
                         android:id="@+id/edit_second"
                         android:padding="4dp"
+                        android:cursorVisible="false"
                         android:layout_width="30dp"
                         android:text="@{timerVM.secondText}"
-                        android:selection="@{timerVM.secondText.length}"
                         android:layout_marginLeft="10dp"
                         android:background="@drawable/number_input_wrap_rounded_box"
                         android:layout_height="30dp"/>
